improved how we configure the bot.
configuration is outsourced to a json file. obviously the bot will not run at all if this file doesn't exist. consult the readme to see what the settings.json file should contain. I've subclassed the bot class with our own that has a configuration property, so all cogs now have access to the config.