altered the display format of the embed slightly.
I feel like the author and contributors will need to be on their own lines. Even with just 2 names there, it was pushing the limit. I added the library because /shrug and changed the links to a masked link. I think besides adding more information if necessary, getting the proper version number in, and figuring out a color, this should be considered feature complete.